requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
gunicorn==22.0.0
idna==3.7
packaging==24.0
pillow==10.3.0
pycparser==2.22
PyMySQL==1.1.0
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.1
